separated pro/player match table js
