+++ public/js/yasp.js
@@ -63,19 +63,37 @@ window.format = function format(input) {
     return (Math.abs(input) < 1000 ? ~~(input) : numeral(input).format('0.0a'));
 }
 window.formatSeconds = function formatSeconds(input) {
-    var absTime = Math.abs(input);
-    var minutes = ~~(absTime / 60);
-    var seconds = pad(~~(absTime % 60), 2);
-    var time = ((input < 0) ? "-" : "");
-    time += minutes + ":" + seconds;
-    return time;
-}
+        var absTime = Math.abs(input);
+        var minutes = ~~(absTime / 60);
+        var seconds = pad(~~(absTime % 60), 2);
+        var time = ((input < 0) ? "-" : "");
+        time += minutes + ":" + seconds;
+        return time;
+    }
+    //extend jquery to serialize form data to JSON
+$.fn.serializeObject = function() {
+    var o = {};
+    var a = this.serializeArray();
+    $.each(a, function() {
+        if (o[this.name] !== undefined) {
+            if (!o[this.name].push) {
+                o[this.name] = [o[this.name]];
+            }
+            o[this.name].push(this.value || '');
+        }
+        else {
+            o[this.name] = this.value || '';
+        }
+    });
+    return o;
+};
 window.tooltips = require('./tooltips.js');
 window.formatHtml = require("./formatHtml.js");
 window.createHistogram = require('./histograms.js');
 window.createCalHeatmap = require('./calheatmap.js');
 window.buildMap = require('./map.js');
 window.playerMatches = require('./playerMatches.js');
+window.proMatches = require('./proMatches.js');
 window.playerTrendsTables = require('./playerTrendsTables.js');
 window.ratingsChart = require('./ratingsChart.js');
 window.statusHandler = require('./statusHandler.js');
