+++ views/player/player.jade
@@ -46,7 +46,7 @@ block content
   include /query
   hr
   block playercontent
-      
+  hr
   table.table.table-condensed
     thead
       tr
