+++ views/professional.jade
@@ -16,5 +16,5 @@ block content
 append footer_assets
   script.
     var pro_matches = !{JSON.stringify(recent)};
-    playerMatches();
+    proMatches(pro_matches);
